+++ src/components/auth/Timer/Timer.tsx
