+++ src/pages/auth/login/index.tsx
@@ -1,0 +1,11 @@
+import LoginForm from "@/components/auth/Form/LoginForm";
+
+const LoginPage = () => {
+  return (
+    <main>
+      <LoginForm />
+    </main>
+  );
+};
+
+export default LoginPage;
