+++ src/pages/auth/signUp/index.tsx
@@ -1,5 +1,5 @@
-import AuthForm from "@/components/signUp/Form/AuthForm";
-import ProfileForm from "@/components/signUp/Form/ProfileForm";
+import AuthForm from "@/components/auth/Form/AuthForm";
+import ProfileForm from "@/components/auth/Form/ProfileForm";
 import signUpStepAtom from "@/recoil/auth/signUp/atom";
 import { SignUpDataType } from "@/types/auth";
 import { FormProvider, useForm } from "react-hook-form";
